feat(playlist): 管理播放列表中的歌曲的相关接口

+++ server/config.js
@@ -74,7 +74,7 @@ const Dir = {
 
 // -- 初始化文件夹 --
 
-if (fs.existsSync(Dir.temp)) {
+if (process.env.TEMP_GUARD !== 'true' && fs.existsSync(Dir.temp)) {
     fs.rmSync(Dir.temp, { recursive: true });
 }
